requirements.txt
discord.py
spotipy
youtube-search-python
pytube
groq
python-dotenv
PyNaCl